Fixing compiler warnings - unboxing parameter to overloaded method, unnecessary typecast
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@805900 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/usermodel/FontDetails.java
@@ -67,7 +67,7 @@ public class FontDetails
      */
     public int getCharWidth( char c )
     {
-        Integer widthInteger = (Integer)(charWidths.get(new Character(c)));
+        Integer widthInteger = charWidths.get(new Character(c));
         if (widthInteger == null && c != 'W') {
             return getCharWidth('W');
         }
